feat: Implement /follow command for spectators
Adds a new chat command, /follow, allowing players in the spectator
force to follow another player's character. This includes the addition
of a safe_wrap_cmd utility for robust command error handling (i.e. a
bug in this implementation will not cause the game to crash, it will
just print out an error).

The implementation is inspired by the /follow command functionality in
Factorio-Biter-Battles.